feat: add query string change on filtration
