+++ src/components/FIlters/StockFilter.scss
