+++ src/components/FIlters/index.ts
@@ -1,0 +1,4 @@
+export { default as CategoryFilter } from "./CategoryFilter";
+export { default as BrandFilter } from "./BrandFilter";
+export { default as StockFilter } from "./StockFilter";
+export { default as PriceFilter } from "./PriceFilter";
