--- src/components/index.ts
@@ -1,3 +1,0 @@
-export {default as CategoryFilter} from './CategoryFilter';
-export {default as BrandFilter} from './BrandFilter';
-export {default as StockFilter} from './StockFilter';
